@@ -5,11 +5,14 @@ var NAVTREE =
     [ "Using the schema compiler", "md__compiler.html", null ],
     [ "Writing a schema", "md__schemas.html", null ],
     [ "Use in C++", "md__cpp_usage.html", null ],
-    [ "Use in Java", "md__java_usage.html", null ],
+    [ "Use in Go", "md__go_usage.html", null ],
+    [ "Use in Java/C-sharp", "md__java_usage.html", null ],
+    [ "Use in Python", "md__python_usage.html", null ],
+    [ "Platform / Language / Feature support", "md__support.html", null ],
     [ "Benchmarks", "md__benchmarks.html", null ],
     [ "FlatBuffers white paper", "md__white_paper.html", null ],
     [ "FlatBuffer Internals", "md__internals.html", null ],
-    [ "Formal Grammar of the schema language", "md__grammar.html", null ]
+    [ "Grammar of the schema language", "md__grammar.html", null ]
   ] ]
 ];
 
