@@ -1,14 +1,17 @@
 var NAVTREEINDEX0 =
 {
 "index.html":[],
-"md__benchmarks.html":[5],
+"md__benchmarks.html":[8],
 "md__building.html":[0],
 "md__compiler.html":[1],
 "md__cpp_usage.html":[3],
-"md__grammar.html":[8],
-"md__internals.html":[7],
-"md__java_usage.html":[4],
+"md__go_usage.html":[4],
+"md__grammar.html":[11],
+"md__internals.html":[10],
+"md__java_usage.html":[5],
+"md__python_usage.html":[6],
 "md__schemas.html":[2],
-"md__white_paper.html":[6],
+"md__support.html":[7],
+"md__white_paper.html":[9],
 "pages.html":[]
 };
