@@ -2,13 +2,14 @@
 
 package MyGame.Example;
 
-import java.nio.*;
-import java.lang.*;
-import java.util.*;
-import flatbuffers.*;
-
-public class Any {
+public final class Any {
+  private Any() { }
   public static final byte NONE = 0;
   public static final byte Monster = 1;
+  public static final byte TestSimpleTableWithEnum = 2;
+
+  private static final String[] names = { "NONE", "Monster", "TestSimpleTableWithEnum", };
+
+  public static String name(int e) { return names[e]; }
 };
 
