@@ -2,14 +2,14 @@
 
 package MyGame.Example;
 
-import java.nio.*;
-import java.lang.*;
-import java.util.*;
-import flatbuffers.*;
+public final class Color {
+  private Color() { }
+  public static final byte Red = 1;
+  public static final byte Green = 2;
+  public static final byte Blue = 8;
 
-public class Color {
-  public static final byte Red = 0;
-  public static final byte Green = 1;
-  public static final byte Blue = 2;
+  private static final String[] names = { "Red", "Green", "", "", "", "", "", "Blue", };
+
+  public static String name(int e) { return names[e - Red]; }
 };
 
